error handling in log maintenance script

+++ utils/log_archive.sh
@@ -13,20 +13,24 @@ echo " - Running Logfile maintenance..."
 
 MAX_FILESIZE_KB=$((5*1024))
 
-for LOG_FILEPATH in $LOGS_DIR/*.log; do
+if [ "$(ls -A $LOGS_DIR)" ]; then
 
-	LOG_FILESIZE_KB=$(($(($(stat $GNU_STAT_FLAG "$LOG_FILEPATH")))/1024))
+	for LOG_FILEPATH in $LOGS_DIR/*.log; do
 
-	if [ "$LOG_FILESIZE_KB" -gt "$MAX_FILESIZE_KB" ]; then
+		LOG_FILESIZE_KB=$(($(($(stat $GNU_STAT_FLAG "$LOG_FILEPATH")))/1024))
 
-		LOG_FILENAME=$(basename -- "$LOG_FILEPATH")
-		CURRENT_DATE_PREFIX=$($GNU_DATE_BIN '+%Y%m%d-%H%M%S')
-		NEW_LOG_FILEPATH="$LOGS_DIR/archive/$CURRENT_DATE_PREFIX-$LOG_FILENAME"
+		if [ "$LOG_FILESIZE_KB" -gt "$MAX_FILESIZE_KB" ]; then
 
-		echo " - Archiving '$LOG_FILENAME' ($LOG_FILESIZE_KB kB)"
+			LOG_FILENAME=$(basename -- "$LOG_FILEPATH")
+			CURRENT_DATE_PREFIX=$($GNU_DATE_BIN '+%Y%m%d-%H%M%S')
+			NEW_LOG_FILEPATH="$LOGS_DIR/archive/$CURRENT_DATE_PREFIX-$LOG_FILENAME"
 
-		mv $LOG_FILEPATH $NEW_LOG_FILEPATH && touch $LOG_FILEPATH && gzip -c $NEW_LOG_FILEPATH > $NEW_LOG_FILEPATH.gz && rm $NEW_LOG_FILEPATH;
+			echo " - Archiving '$LOG_FILENAME' ($LOG_FILESIZE_KB kB)"
 
-	fi
+			mv $LOG_FILEPATH $NEW_LOG_FILEPATH && touch $LOG_FILEPATH && gzip -c $NEW_LOG_FILEPATH > $NEW_LOG_FILEPATH.gz && rm $NEW_LOG_FILEPATH;
 
-done
+		fi
+
+	done
+
+fi
